Added a security check in my config to try to prevent insecure execution
There's a code vulnerability patched in 8.1.1365 that allows your system
to be hacked through a code exploit.  In an attempt to prevent this, I
added config that will quit vim before we get to load a file if the
version of Vim doesn't include the patch.

Details on the vulnerability:
https://nvd.nist.gov/vuln/detail/CVE-2019-12735

It's a bit crude, but it seems to work.  It's effectively enforcing a
minimum Vim version to execute.

Also, disable the modeline functionality, as an additional precaution.
I'm unsure if this will have any issues with plugins, but I can deal
with that if it does.